files | fix read write (IB)

+++ metro-system-CLI/Station.cpp
@@ -170,13 +170,13 @@ bool Station::deserialize(std::istream& is) {
     name = readString(is);
     idCounter = id;
 
-    int sizeOfLinesIdsVector;
+    size_t sizeOfLinesIdsVector;
     if (!is.read(reinterpret_cast<char*>(&sizeOfLinesIdsVector), sizeof(sizeOfLinesIdsVector)))
         return false;
     linesIds.resize(sizeOfLinesIdsVector);
     is.read(reinterpret_cast<char*>(linesIds.data()), sizeOfLinesIdsVector * sizeof(int));
 
-    int sizeOflinkedStationIdsVector;
+    size_t sizeOflinkedStationIdsVector;
     if (!is.read(reinterpret_cast<char*>(&sizeOflinkedStationIdsVector), sizeof(sizeOflinkedStationIdsVector)))
         return false;
     linkedStationIds.resize(sizeOflinkedStationIdsVector);
